Time: 120 ms (86.26%) | Memory: 63 MB (49.76%) - LeetSync

+++ problems/2616-maximal-score-after-applying-k-operations/maximal-score-after-applying-k-operations.java
@@ -1,14 +1,14 @@
 class Solution {
     public long maxKelements(int[] nums, int k) {
-        PriorityQueue<Long> pq = new PriorityQueue<>((a,b)->Long.compare(b,a));
+        PriorityQueue<Integer> pq = new PriorityQueue<>((a,b)->Integer.compare(b,a));
         for(int value:nums){
-            pq.add((long)value);
+            pq.add(value);
         }
         long score=0;
         while(k>0){
-            long value = pq.poll();
+            int value = pq.poll();
             score+=value;
-            pq.add((long)Math.ceil(value/3.0));
+            pq.add((int)Math.ceil(value/3.0));
             k--;
         }
         return score;
